+++ Client/client.cpp
@@ -18,16 +18,18 @@
 -- NOTES:       Server class methods for creating and using a client for the FTP program
 --------------------------------------------------------------------------------------------------*/
 
-#include "../A1/client.h"
-#include "../A1/server.h"
-#include "../A1/socketwrappers.h"
+#include "../Library/socketwrappers.h"
+#include "../Library/library.h"
+#include "../Library/client.h"
+#include "../Library/server.h"
+
 
 #define PATH "./Client_files/"
 
 using namespace std;
 
-void RecvFile(int socket, char* filename);
-void SendFile(int socket, char *filename);
+//void RecvFile(int socket, char* filename);
+//void SendFile(int socket, char *filename);
 bool isValidFile(char *cfilename);
 bool isCommand(string strcommand, int &command);
 char *ParseString(string str);
@@ -90,7 +92,7 @@ int main (int argc, char *argv[]) {
 		//create a command
 		cmd = CreateCmd(command, cfilename);
 		//send the command
-		commandConnection->SendCmd(cmd);
+		SendCmd(commandConnection->GetSocket(), cmd);
 		//send or receive the file
 		if(cmd.type == SEND) {
 			printf("sending file\n");
@@ -197,118 +199,4 @@ bool isValidFile(char *cfilename) {
 	return true;
 }
 
-/*-----------------------------------------------------------------------------------------------
--- FUNCTION:   SendFile
---
--- DATE:       Oct 2, 2017
---
--- DESIGNER:   Aing Ragunathan
---
--- PROGRAMMER: Aing Ragunathan
---
--- INTERFACE:  void SendFile(int socket, char *filename) 
---
--- PARAMETER:  	int socket 		- socket to send the data into
---				char *filename	- file to send through the socket
---
--- RETURNS:    void
---
--- NOTES:      Sends an entire file through the socket
------------------------------------------------------------------------------------------------ */
-void SendFile(int socket, char *filename) {
-	FILE *file;
-	char buffer[BUFLEN];
-	int bytesRead, bytesSent, seqNum;	
-	Packet packet;
 
-	if((file = fopen(filename, "rb")) == NULL) {
-		perror("file doesn't exist\n");
-		return;
-	}
-
-	memset(buffer, '\0', BUFLEN);
-	memset(packet.Data, '\0', BUFLEN);
-
-	while((bytesRead = fread(buffer, 1, sizeof(buffer), file)) != 0) {
-		//calculate the sequence number for the next packet (move to end of loop)
-		seqNum = bytesRead + packet.SeqNum;
-
-		//Send the DATA packet
-		packet = CreatePacket(DATA, seqNum, buffer, 0, 0);
-		if((bytesSent = write(socket, &packet, sizeof(packet))) == -1) {
-			perror("Error writing to socket DATA: ");
-			return;
-		} 
-
-		memset(buffer, '\0', BUFLEN);
-		
-		//send the EOT packet: End of transmission
-		if(bytesRead < BUFLEN) {
-			packet = CreatePacket(EOT, seqNum, buffer, 0, 0);	
-			if((bytesSent = write(socket, &packet, sizeof(packet))) == -1) {
-				perror("Error writing to socket EOT: ");
-				return;
-			}
-		}
-	}
-	
-	fclose(file);
-}
-
-/*-----------------------------------------------------------------------------------------------
--- FUNCTION:   RecvFile
---
--- DATE:       Oct 2, 2017
---
--- DESIGNER:   Aing Ragunathan
---
--- PROGRAMMER: Aing Ragunathan
---
--- INTERFACE:  void RecvFile(int socket, char* filename)
---
--- PARAMETER:  	int socket 		- socket to send the data into
---				char *filename	- file to send through the socket
---
--- RETURNS:    void
---
--- NOTES:      Receives an entire file through the socket
------------------------------------------------------------------------------------------------ */
-void RecvFile(int socket, char* filename) {
-	FILE *file;
-	char buffer[BUFLEN];
-	int bytesRecv, writeCount = 0;
-	Packet packet;
-
-	//open file to write in binary
-	if((file = fopen(filename, "wb")) == NULL) {
-		printf("file failed to open: %s\n", filename);
-		return;
-	}	
-
-	truncate(filename, 0);
-
-	while(1) {
-		bzero(buffer, sizeof(buffer));
-
-		//receive the packet
-		if((bytesRecv = read(socket, &packet, sizeof(packet))) < 0) {
-			perror("Error receiving from socket");
-			return;
-		} else {
-			//check the packet type and treat accordingly
-			if(packet.Type == DATA) {
-				PrintPacket(packet);	//print content of file
-			
-				if((writeCount = fwrite(packet.Data, 1, strlen(packet.Data), file)) < 0) {
-					perror("Write failed");
-					return;
-				}
-
-			} else if(packet.Type == EOT) {
-				printf("Type: EOT\n");
-				fclose(file);
-				return;
-			}
-		}
-	}
-}
