+++ Client/makefile
@@ -1,2 +1,2 @@
-client: client.cpp ../A1/server.cpp ../A1/client.cpp ../A1/socketwrappers.cpp
-	g++ -g -Wall -o client client.cpp ../A1/server.cpp ../A1/client.cpp ../A1/socketwrappers.cpp
+client: client.cpp ../Library/server.cpp ../Library/client.cpp ../Library/socketwrappers.cpp ../Library/library.cpp
+	g++ -g -Wall -o client client.cpp ../Library/server.cpp ../Library/client.cpp ../Library/socketwrappers.cpp ../Library/library.cpp
