+++ Library/client.h
@@ -2,6 +2,7 @@
 #define CLIENT_H
 
 #include "socketwrappers.h"
+#include "library.h"
 #include <netinet/in.h>
 #include <iostream> 
 #include <strings.h>
