+++ Library/socketwrappers.cpp
@@ -319,106 +319,6 @@ int RecvMsg(int sockfd, char* buffer) {
 	return bytesRecv;
 }
 
-/*-----------------------------------------------------------------------------------------------
--- FUNCTION:   RecvCmd
---
--- DATE:       Oct 2, 2017
---
--- DESIGNER:   Aing Ragunathan
---
--- PROGRAMMER: Aing Ragunathan
---
--- INTERFACE:  Cmd RecvCmd(int sockfd)
---
--- PARAMETER:  	int sockfd 		- file descriptor to connect with
---
--- RETURNS:    return a Cmd object containing a command type and filename
---
--- NOTES:      wrapper function for receiving a Cmd object
------------------------------------------------------------------------------------------------ */
-Cmd RecvCmd(int sockfd) {
-	Cmd cmd;
-	int bytesRecv;
-
-	if((bytesRecv = recv(sockfd, &cmd, sizeof(Cmd), 0)) == -1) {
-		perror("RecvCmd Failed");
-	} else if(bytesRecv == 0) {
-		printf("Connection ended\n");
-		cmd = CreateCmd(0, NULL);
-	}
-
-    printf("RecvCmd: %d %s\n", cmd.type, cmd.filename);
-
-	return cmd;
-}
-
-/*-----------------------------------------------------------------------------------------------
--- FUNCTION:   CreateCmd
---
--- DATE:       Oct 2, 2017
---
--- DESIGNER:   Aing Ragunathan
---
--- PROGRAMMER: Aing Ragunathan
---
--- INTERFACE:  Cmd CreateCmd(int type, char *filename) 
---
--- PARAMETER:  	int type 		- the type of command to attach to the Cmd object
---				char *filename	- the filename to attach to the Cmd object
---
--- RETURNS:    return a Cmd object containing a command type and filename
---
--- NOTES:      wrapper function for creating a Cmd object
------------------------------------------------------------------------------------------------ */
-Cmd CreateCmd(int type, char *filename) {
-	Cmd cmd;
-
-	cmd.type = type;
-    strcpy(cmd.filename, filename);
-
-	return cmd;	
-}
-
-/*-----------------------------------------------------------------------------------------------
--- FUNCTION:   CreatePacket
---
--- DATE:       No 11, 2017
---
--- DESIGNER:   Aing Ragunathan
---
--- PROGRAMMER: Aing Ragunathan
---
--- INTERFACE:  Packet CreatePacket(int type, int SeqNum, char *PayloadLen, int WindowSize, int AckNum) 
---
--- PARAMETER:  	
---
--- RETURNS:   
---
--- NOTES:    
------------------------------------------------------------------------------------------------ */
-Packet CreatePacket(int type, int SeqNum, char data[BUFLEN], int WindowSize, int AckNum) {
-	Packet packet;
-
-	packet.Type = type;
-	packet.SeqNum = SeqNum; 
-    	strcpy(packet.Data, data);
-	packet.WindowSize = WindowSize;
-	packet.AckNum = AckNum;	
-
-	return packet;	
-}
-
-//print content of the file
-void PrintPacket(Packet packet) {
-	printf("Type: Data\n");
-        printf("SeqNum: %d\n", packet.SeqNum);
-	printf("WinSize: %d\n", packet.WindowSize);
-	printf("AckNum: %d\n", packet.AckNum);
-	printf("Payload: %s", packet.Data);
-}
-
-
-
