+++ Server/makefile
@@ -1,2 +1,2 @@
-server: server.cpp ../A1/server.cpp ../A1/client.cpp ../A1/socketwrappers.cpp
-	g++ -g -Wall -o server server.cpp ../A1/server.cpp ../A1/client.cpp ../A1/socketwrappers.cpp
+server: server.cpp ../Library/server.cpp ../Library/client.cpp ../Library/socketwrappers.cpp ../Library/library.cpp
+	g++ -g -Wall -o server server.cpp ../Library/server.cpp ../Library/client.cpp ../Library/socketwrappers.cpp ../Library/library.cpp
